src/odl_kernel/types/__init__.py
# Copyright (c) 2026 Centillion System, Inc. All rights reserved.
#
# This software is licensed under the Business Source License 1.1 (BSL 1.1).
# For full license terms, see the LICENSE file in the root directory.
#
# This software is "Source Available" but contains certain restrictions on
# commercial use and managed service provision. 
# Usage by organizations with gross revenue exceeding $10M USD requires 
# a commercial license.

from .enums import (
    JobStatus,
    LifecycleStatus,
    BusinessResult,
)

from .entities import (
    Job,
    ProcessNode,
    ContextSchema,
)

from .structs import (
    # Input
    JobSnapshot,
    KernelEvent,
    KernelEventType,
    InterventionIntent,
    
    # Output
    AnalysisResult,
    RuntimeCommand,
    CommandType,
    JobUpdate,
)

# Re-export core ODL types for convenience
# これにより odl_kernel.types.OpCode のようにアクセス可能にする
from odl.types import (
    OpCode,
    NodeType,
    IrComponent,
    WiringObject,
    WorkerMode,
)

__all__ = [
    # Enums
    "JobStatus",
    "LifecycleStatus",
    "BusinessResult",
    
    # Entities
    "Job",
    "ProcessNode",
    "ContextSchema",
    
    # Structs (Input/Output)
    "JobSnapshot",
    "KernelEvent",
    "KernelEventType",
    "InterventionIntent",
    "AnalysisResult",
    "RuntimeCommand",
    "CommandType",
    "JobUpdate",

    # ODL Standard Types
    "OpCode",
    "NodeType",
    "IrComponent",
    "WiringObject",
    "WorkerMode",
]